@@ -1,6 +1,6 @@
 package cc.brainbook.android.headerdecoration.sample;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 import java.util.Arrays;
