@@ -1,7 +1,7 @@
 package cc.brainbook.android.headerdecoration.util;
 
 import android.graphics.Rect;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.View;
 
 import static android.view.ViewGroup.LayoutParams;
